# sha256: use freelist to avoid allocations